refactor: swagger를 위해 분리한 인터페이스 클래스명 Api에서 Swagger로 변경
